refactor: rename adapter to provider and move to providers folder
