+++ packages/solana/src/connectors/walletConnectConnector.ts
@@ -5,11 +5,10 @@ import { OptionsController } from '@web3modal/core'
 import { SolStoreUtil } from '../utils/scaffold/index.js'
 import { UniversalProviderFactory } from './universalProvider.js'
 import { BaseConnector } from './baseConnector.js'
-import type { Connector } from './baseConnector.js'
 
 import type UniversalProvider from '@walletconnect/universal-provider'
 
-import type { Chain, AnyTransaction } from '../utils/scaffold/SolanaTypesUtil.js'
+import type { Chain, AnyTransaction, Provider } from '../utils/scaffold/SolanaTypesUtil.js'
 import {
   getChainsFromChainId,
   getDefaultChainFromSession,
@@ -24,7 +23,7 @@ export interface WalletConnectAppMetadata {
   icons: string[]
 }
 
-export class WalletConnectConnector extends BaseConnector implements Connector {
+export class WalletConnectConnector extends BaseConnector implements Provider {
   id = 'WalletConnect'
   name = 'WalletConnect'
   ready = true
@@ -84,6 +83,8 @@ export class WalletConnectConnector extends BaseConnector implements Connector {
     return base58.decode(signedMessage.signature)
   }
 
+  public isConnected = () => true
+
   public async signTransaction<T extends AnyTransaction>(transaction: T) {
     const serializedTransaction = this.serializeTransaction(transaction)
 
@@ -106,6 +107,8 @@ export class WalletConnectConnector extends BaseConnector implements Connector {
     sendOptions?: SendOptions
   ) {
     const serializedTransaction = this.serializeTransaction(transaction)
+
+    console.log('serialized transaction', serializedTransaction)
 
     const result = await this.request('solana_signAndSendTransaction', {
       transaction: serializedTransaction,
