+++ packages/solana/src/providers/WalletStandardProvider/errors.ts
@@ -1,4 +1,4 @@
-export class StandardWalletFeatureNotSupportedError extends Error {
+export class WalletStandardFeatureNotSupportedError extends Error {
   constructor(feature: string) {
     super(`The wallet does not support the "${feature}" feature`)
   }
